Refactor GUI to use tabs
